Adapt to moved PREFIX and move some files
