+++ src/controllers/IframeController.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from 'react-router-dom';
 import { useCurrentStep } from '../routes/utils';
 import { useStoreDispatch, useStoreActions, useFlatSequence } from '../store/store';
 import { WebsiteComponent } from '../parser/types';
-import { PREFIX as BASE_PREFIX } from '../components/Prefix';
+import { PREFIX as BASE_PREFIX } from '../utils/Prefix';
 
 const PREFIX = '@REVISIT_COMMS';
