+++ src/controllers/MarkdownController.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import ReactMarkdownWrapper from '../components/ReactMarkdownWrapper';
 import { MarkdownComponent } from '../parser/types';
-import { PREFIX } from '../components/Prefix';
+import { PREFIX } from '../utils/Prefix';
 
 export default function MarkdownController({ currentConfig }: { currentConfig: MarkdownComponent; }) {
   const [importedText, setImportedText] = useState<string | null>(null);
